review processes; add proc-macro

+++ src/process/pid.rs
@@ -1,7 +1,7 @@
 use crate::declare_constants;
 declare_constants!(
     pub usize,
-    MAX_PID_COUNT = u16::MAX, "The maximal count of unique processes in the system";
+    MAX_PID_COUNT = u16::MAX as usize, "The maximal count of unique processes in the system";
 );
 pub struct PidBitmap {
     map: usize,
